Fix tarantoolctl cat/play premature stop with --to
Stop a file processing loop only when it is guaranteed that we will not
find a record that match user-provided filters later in this file. If
--replica R is provided one time and we're meet a record from R with a
LSN equal or above of a --to value, we'll stop the loop. Otherwise (no
--replica, several --replica arguments) a file will be read until an end
even if --to is provided.

Fixes #3827.

(cherry picked from commit 70a8342fbb43a9d72c1e035c4d41551b729dcb9c)