test: fix test for optional SQL_BIND in iproto
Added necessary cleanup, because the test fails when both memtx and
vinyl configurations are run on the same test-run worker.

Aside of that disabled the test on vinyl, because it is engine-agnostic
and only checks iproto.

This is the fix for the commit 7676b2b1 ('sql: make SQL_BIND optional in
an iproto request').

Follows up #4077.

(cherry picked from commit f6520860c4e9a4030e02acaf17c53f581fbd4701)