sql: disallow text values participate in sum() aggregate
It is obvious that we can't add string with number except the case when
string is a number literal in quotes (aka '123' or '0.5'). Before this
patch values which can't be converted to numbers were just skipped.
Now error is raised. This patch also removes sql_value_numeric_type()
function since it is not used anymore: instead we invoke
mem_apply_numeric_type().