sql: remove "PRAGMA case_sensitive_like"
According to ANSI, LIKE should match characters taking into account
collations of arguments, and this is done in the next patch. In turn,
this patch makes LIKE be always case sensitive and erases
case_sensitive_like pragma. Alongside with it, code related to no-case
LIKE optimization is removed as well.

Part of #3589