swim: drop incarnation_inc parameter from update() routines
Update_addr and update_payload need to increment member's
incarnation when it is self. For that they used a special
parameter incarnation_inc set in 1 for self and in 0 for others.

It was used to encapsulate incarnation update + event scheduling
on member attribute updates, but on the other hand it broke
another encapsulation level - there should not be exceptions for
'self' in these functions.

This patch makes incarnation increment explicit in the places
where 'self' is updated.