Refactor tarantool search and loader functions
This commit refactor the way how tarantool handles searching and loading
modules that is out of scope of default lua loaders.

Dublicate code was remove. Internal function api was unified, so adding
additional loader are easier and more flexible now. Amount of hardcoded
values was reduces.

It adds package.get_scriptpath function, which uses debug module
to get current path of the caller's source file

It inserts two additional loaders to package, which is used to look for
plain lua modules and libraries inside application directory